+++ config/greenlight.php
@@ -1,0 +1,6 @@
+<?php
+
+return [
+    'compatibility' => env('GREENLIGHT_COMPATIBILITY',false),
+    'base' => env('GREENLIGHT_PATH',"b"),
+];